\name{modulePreservation}
\alias{modulePreservation}
\title{Calculates pairwise between reference and test networks}
\usage{
  modulePreservation(expressionSets = NULL,
    adjacencySets = NULL, referenceSets, testSets,
    moduleLabels, nPermutations = 200, verbose = TRUE,
    save = TRUE, permutedFile = "permutedStatistics.rda")
}
\arguments{
  \item{expressionSets}{}

  \item{adjacencySets}{}

  \item{referenceSets}{}

  \item{testSets}{}

  \item{nPermutations}{}

  \item{moduleLabels}{}

  \item{verbose}{}

  \item{savePermuted}{}

  \item{permutedFile}{}
}
\value{
  a
}
\description{
  Description goes here
}

